fix: speed up /news with SSR prefetch and longer image cache TTL
Two bottlenecks were causing the /news page to feel slow:

1. Image optimizer cache too short
   - Default minimumCacheTTL=60s meant every news thumbnail re-fetched
     from GCS and re-encoded after a minute. Set to 30 days since the
     thumbnail URL identifies content.
   - Add AVIF→WebP format ladder (~30-50% payload cut).

2. Client-side data waterfall
   - Page was "use client" with useSuspenseQuery firing only after
     hydration. User saw skeleton until /news/list responded.
   - Split into server page (SSR prefetch + dehydrate) and client
     page-client.tsx (pagination + suspense). HydrationBoundary feeds
     prefetched data so first paint already has the cards.
   - Removed "use client" from news.query.ts so the queryOptions
     factory is server-importable; hooks remain usable from client
     components without a directive (consumer is what matters).
