+++ next.config.mjs
@@ -19,6 +19,10 @@ const nextConfig = {
 				hostname: "storage.googleapis.com",
 			},
 		],
+		/* 30일 캐시 — news 썸네일 URL은 idx별로 안정적이라 길게 잡아도 안전 */
+		minimumCacheTTL: 60 * 60 * 24 * 30,
+		/* AVIF 우선 → WebP 폴백 — 페이로드 30~50% 감소 */
+		formats: ["image/avif", "image/webp"],
 	},
 
 	async headers() {
